[msedgeredirect] Fix package metadata
Amend the copyright field. You have added the license name in there.

Remove the comments from the chocolateyInstall.ps1

Add a line to the top of the LICENSE.txt that points to the official
license (ie. what is in the licenseUrl field) in the format From:
&lt;LICENSE URL&gt;.

Explain the reason for more than 5 AV detections on the
msedgeredirectinstall.exe and point to an official explanation
in the description field?